rename confusing basic_plugin plugins
