+++ plugins/basic_plugin_watcher/java.json
