+++ plugins/basic_plugin_watcher/js.json
